I added a parallel G_build to libscf_solver.  Also, proc 0 does all chkpt rd/wt followed by a raw_bcast in libscf_solver and libmints.
git-svn-id: https://sirius.chem.vt.edu/svn/psi4/trunk@939 6a214465-e6f0-41bc-91e6-de896bb96b5e

+++ src/lib/libmints/molecule.cc
@@ -844,7 +844,7 @@ void Molecule::init_with_chkpt(shared_ptr<Chkpt> chkpt)
             geom = block_matrix(natoms, 3);
         }
 
-        Communicator::world->raw_bcast(&(zvals[0]), natoms*sizeof(int), 0);
+        Communicator::world->raw_bcast(&(zvals[0]), natoms*sizeof(double), 0);
         Communicator::world->raw_bcast(&(geom[0][0]), natoms*3*sizeof(double), 0);
     }
